Created an images folder in assets and moved all images into it
